requirements.txt
alabaster==0.7.12
asgiref==3.3.4
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.5
dj-database-url==0.5.0
Django==3.2.3
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docutils==0.17.1
gunicorn==20.1.0
idna==2.10
imagesize==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
snowballstemmer==2.1.0
Sphinx==4.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
whitenoise==5.2.0
